requirements.txt
allure-pytest==2.8.21
allure-python-commons==2.8.21
attrs==20.3.0
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
